requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
Faker==27.0.0
Markdown==3.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
validate-docbr==1.10.0
